优化订单处理逻辑，调整数据模型，更新配置文件，修复结账服务中的错误

+++ app/order/biz/model/order_item.go
@@ -6,8 +6,8 @@ type OrderItem struct {
 	gorm.Model
 	ProductID uint32  `json:"productId"`
 	OrderID   string  `json:"orderId" gorm:"type:varchar(100)"`
-	Quantity  uint32  `json:"quantity"`
-	Cost      float64 `json:"cost" gorm:"type:decimal(10,2)"`
+	Quantity  int32  `json:"quantity"`
+	Cost      float32 `json:"cost" gorm:"type:decimal(10,2)"`
 }
 
 func (OrderItem) TableName() string {
